Move library function into quantaq_py/toolkit
