+++ quantaq_py/__init__.py
@@ -1,9 +1,9 @@
-from quantaq_py.concat import concat_files
-from quantaq_py.merge import merge_files
-from quantaq_py.resample import resample_dataframe
-from quantaq_py.flag import flag_dataframe, echo_flag_table, flag_summary
-from quantaq_py.expunge import expunge_dataframe
-from quantaq_py.munge import clean_file
+from quantaq_py.toolkit.concat import concat_files
+from quantaq_py.toolkit.merge import merge_files
+from quantaq_py.toolkit.resample import resample_dataframe
+from quantaq_py.toolkit.flag import flag_dataframe, echo_flag_table, flag_summary
+from quantaq_py.toolkit.expunge import expunge_dataframe
+from quantaq_py.toolkit.munge import clean_file
 
 
 __all__ = [
