+++ tests/test_resample.py
@@ -10,7 +10,7 @@ import pandas as pd
 from click.testing import CliRunner
 from loguru import logger
 
-from quantaq_py.resample import resample_dataframe
+from quantaq_py.toolkit.resample import resample_dataframe
 from quantaq_py.cli import resample_command
 from quantaq_py.utilities import safe_load
